move condalib into conda proper
add conda.base module

update common readme
