--- condalib/common/compat.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import, division, print_function
-from logging import getLogger
-
-log = getLogger(__name__)
-
-from auxlib.compat import (iteritems, with_metaclass, itervalues, string_types,  # NOQA
-                           primitive_types, text_type, odict, StringIO, isiterable)  # NOQA
