--- condalib/config/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import, division, print_function
-from logging import getLogger
-
-log = getLogger(__name__)
